To make the layering more clear, factor out dispatch_method stuff.
The context module is the stuff needed to support dispatch methods on
a context class. The dispatch module contains just the dispatch function
stuff.
